requirements.txt
opencv-python
torch
torchvision
Pillow
fastapi
uvicorn
ultralytics
numpy
scikit-learn
requests